+++ packages/waggle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+* Enable cors in DataServer only for development and end to end testing
+
 ## [4.3.8]
 
 ### Fixed